[mc_control] Move many FSM functionalities to MCController
