+++ debian/libmc-rtc-dev.install
@@ -5,6 +5,7 @@ usr/include/mc_rbdyn/*
 usr/include/mc_robots/*
 usr/include/mc_rtc/gui/*
 usr/include/mc_rtc/log/*
+usr/include/mc_rtc/clock.h
 usr/include/mc_rtc/config.h
 usr/include/mc_rtc/Configuration.h
 usr/include/mc_rtc/ConfigurationHelpers.h
